+++ frontend/src/features/admin/admin-areas/interfaces/addAdminArea.ts
@@ -1,0 +1,4 @@
+/** required interface for adding a new administrative area */
+export interface IAddAdminArea {
+  name: string;
+}
