+++ frontend/src/features/admin/admin-areas/interfaces/adminAreaFilter.ts
@@ -1,0 +1,8 @@
+/** interface containing filterable values for administrative areas */
+export interface IAdminAreaFilter {
+  name?: string;
+  id?: number;
+  page?: number;
+  quantity?: number;
+  sort?: string | string[];
+}
