+++ frontend/src/features/admin/admin-areas/interfaces/index.ts
@@ -1,0 +1,4 @@
+export * from './addAdminArea';
+export * from './adminAreaFilter';
+export * from './administrativeArea';
+export * from './apiAdminArea';
